Frontend/requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.1
protobuf==3.20.3